CodigosPythonAWSIoT_MQTT_Paho/requirements.txt
altgraph==0.16.1
awscli==1.16.225
AWSIoTPythonSDK==1.4.7
boto3==1.9.215
botocore==1.12.215
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
coloredlogs==10.0
docutils==0.15.2
enum34==1.1.6
future==0.17.1
humanfriendly==4.18
idna==2.8
jmespath==0.9.4
macholib==1.11
paho-mqtt==1.4.0
pefile==2019.4.18
pyasn1==0.4.6
PyInstaller==3.5
pyreadline==2.1
pyserial==3.4
python-dateutil==2.8.0
pywin32-ctypes==0.2.0
PyYAML==5.1.2
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
schedule==0.6.0
six==1.12.0
urllib3==1.25.3
watchgod==0.4
